% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yomi.R
\name{moji_to_yomi}
\alias{moji_to_yomi}
\title{Find out how to read}
\usage{
moji_to_yomi(x)
}
\arguments{
\item{x}{kanji character}
}
\description{
Find out how to read
}
\examples{
moji_to_yomi("\U74DC")
moji_to_yomi("\U3436")
}
