% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{sample_moji}
\alias{sample_moji}
\title{Random Sample for Kanji}
\usage{
sample_moji(size)
}
\arguments{
\item{size}{a positive number, the number of items to choose from.}
}
\description{
Random Sample for Kanji
}
\examples{
sample_moji(1)
}
