% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{is_jyouyou_kanji}
\alias{is_jyouyou_kanji}
\alias{is_jinmeiyou_kanji}
\alias{kakusu}
\alias{busyu}
\title{Check how Kanji characters are used}
\usage{
is_jyouyou_kanji(x)

is_jinmeiyou_kanji(x)

kakusu(x)

busyu(x)
}
\arguments{
\item{x}{kanji character}
}
\description{
Check how Kanji characters are used
}
\examples{
is_jyouyou_kanji("\U74DC")
is_jyouyou_kanji("\U6A39")
is_jyouyou_kanji(c("\u8af8", "\u884c", "\u7121", "\u5e38"))

is_jinmeiyou_kanji("\U74DC")
is_jinmeiyou_kanji(c("\u590f", "\u76ee", "\u6f31", "\u77f3"))

kakusu("\U74DC")
busyu("\U74DC")
}
