% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{moji_information}
\alias{moji_information}
\title{List for moji information}
\usage{
moji_information(x)
}
\arguments{
\item{x}{kanji character}
}
\description{
List for moji information
}
\examples{
moji_information("\U74dc")
}
