% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{is_jyouyou_kanji}
\alias{is_jyouyou_kanji}
\alias{is_jinmeiyou_kanji}
\title{Check how Kanji characters are used}
\usage{
is_jyouyou_kanji(x)

is_jinmeiyou_kanji(x)
}
\arguments{
\item{x}{kanji character}
}
\description{
Check how Kanji characters are used
}
\examples{
is_jyouyou_kanji("\U74dc")
is_jyouyou_kanji("\U6A39")

is_jinmeiyou_kanji("\U74dc")
}
